requirements.txt
mitmproxy
adblockparser
pyre2